merger: Keep redundant cherry-pick commits
In normal git usage, cherry-picking a commit that has already been
applied doesn't do anything, or cherry-picking an empty commit causes
git to exit with an error to let the user decide what they want to do.
However, this doesn't match the behavior of merges where the commit is
simply skipped if it has already been applied, or is kept if empty.

To fix this, add the --keep-redundant-commit option to `git cherry-pick`
to make git add commit when cherry-picking even when it is empty for
either reason. Then, after the cherry-pick, check if the commit new
commit is empty and if so back it out if the original commit wasn't
empty.

This two step process is necessary because git doesn't have any options
to simply skip cherry-pick commits that have already been applied to the
tree.

Remove commits that have already been applied is particularly important
in a "deploy" pipeline triggered by a Gerrit "change-merged" event,
since the scheduler will try to cherry-pick the change on top of the
commit that just merged. Without this option, the cherry-pick will fail
and the deploy pipeline will fail with a MERGE_CONFICT.

Change-Id: I326ba49e2268197662d11fd79e46f3c020675f21