Update reconfig event ltime on (smart) reconfig
Make sure to update the last reconfig event ltime in the layout state
during a (smart) reconfig. This is important in order to unblock
pipelines when a tenant reconfig event is lost.

So far the last reconfig event ltime was passed as -1, but wasn't set in
the layout state since the ltime is not allowed to go backward.

TODO: needs tests
Change-Id: Iab04a962abbfbe901c22e4d5f1d484e3f53b0d33