dnm: Store generic trigger event in queue item
The event that's currently stored as part of the queue item is not
sharded. This means that we can see Zookeeper disconnects when the queue
item data exceeds the max. Znode size of 1MB.

This change tests what would break if we would just store the generic
trigger event fields and no driver specific data.

Change-Id: I13577498e55fd4bb189679836219dea4dc5729fc