Add GitHub pipeline trigger requirements
This mimics a useful feature of the Gerrit driver and allows users
to configure pipelines that trigger on events but only if certain
conditions of the PR are met.

One of the 6 variations here, "require-status", already exists but
is undocumented.

Some documentation typos are also corrected, and the shared classes
in the GitHub driver are updated to achieve just a bit more
consistency.

Change-Id: I4b6dd8c970691b1e74ffd5a96c2be4b8075f1a87