Optimize layout generation with many job variants
Systems with very large numbers of job variants can suffer slow
performance due to inefficiencies in layout generation.  This change
updates two locations to be more efficient:

1) When adding a job variant, we check to make sure that the variant
we are adding is from the same repo as all the previous variants.
Due to the transitive property of this check, we only need to compare
the new variant to one of the prior variants, not all of them.

2) When validating references on a job variant, we created 3 list
comprehensions with all of the variant's parents which matched certain
criteria.  Since most variants match all three criteria, if we
short-circuit this check as soon as all three criteria have been
satisfied, we can complete the process much sooner.

Change-Id: Iaa9cd663581b3b888fb1ee95c0c25a3484d3e828