Set layout state event ltime in delete-pipeline-state
The delete-pipeline-state command updates the layout state in order
to force schedulers to update their local layout (essentially perform
a local-only reconfiguration).  In doing so, it sets the last event
ltime to -1.  This is reasonable for initializing a new system, but
in an existing system, when an event arrives at the tenant trigger
event queue it is assigned the last reconfiguration event ltime seen
by that trigger event queue.  Later, when a scheduler processes such
a trigger event after the delete-pipeline-state command has run, it
will refuse to handle the event since it arrived much later than
its local layout state.

This must then be corrected manually by the operator by forcing a
tenant reconfiguration.  This means that the system essentially suffers
the delay of two sequential reconfigurations before it can proceed.

To correct this, set the last event ltime for the layout state to
the ltime of the layout state itself.  This means that once a scheduler
has updated its local layout, it can proceed in processing old events.

Change-Id: I66e798adbbdd55ff1beb1ecee39c7f5a5351fc4b