Don't discard all cat job results in case of error
So far we've aborted all cat jobs when any of the cat jobs failed.
However, since we ignore those exceptions anyways unless we are
validating the tenant configuration, we should continue processing the
rest of the job results that haven't failed.

If an early cat job failed we might otherwise not load configuration
from repositories and branches even if those cat jobs were successful.

Change-Id: I34f2a23641de9138b1e887df86ae2602ca190277