Cache JWT well-known url in zuul-web
If the keys_url is not supplied to a JWT authenticator, Zuul will
look up the url based on the well-known config url for the authenticator.
Since that is not typically expected to change, we can cache that value
and avoid repeating the lookup for every authenticated request we receive.

This can significantly reduce the outbound network traffic from zuul-web.

Change-Id: Ibbce7c8c90f9ccbfdcf41d6635162681a8fe3319