Deduplicate circular dep jobs in check
If a cycle of dependencies in a dependent pipeline meet certain
conditions and share some jobs, we deduplicate those jobs (i.e.,
we run only one build of the job and share it between multiple
items simultaneously).

This makes the same happen for independent pipelines.  This is
accomplished by searching for identical bundles within the pipeline
and deduplicating the job across bundles.  The deduplication is
no different than dependent pipelines -- we simply store a reference
to the actual build in ZK and use the same Build object in Python.

In dependent pipelines, bundle queue items are removed
simultaneously (they all report at the same time).  However, in
independent pipelines, they may report at different times.  If
a queue item is dequeued while one of its builds is referenced by
a different queue item, then the second queue item will begin to
have errors loading information from ZK.

To deal with this, we utlilize a property of how we store items in
ZK: the QueueItem itself is stored underneath the pipeline object
with no reference to its place in the queue.  Separately there are
queue lists which reference the QueueItem zobjects by path.

We will continue to remove the QueueItem from the Queue, but we will
leave it in place in ZK if it is possible that it may be referenced
by other queue items.  Then we will rely more heavily on a garbage
collection system to remove the QueueItems once they, and their builds,
are no longer referenced by any item in the pipeline.

The existing garbage collection happens after loading the pipeline
state from ZK, but this change moves it to the end of pipeline
processing so that we are more likely to write out empty pipelines
at the end of the process once all items are removed.

Change-Id: I7fc897088d720f483e320431fefaa06425a33c86