web: add dark mode
This adds a config option to enable dark mode that
is persistented through the use of localStorage.

Upon loading we need to modify the html tag in
DOM but we also update the darkMode variable in
the state so that it can be used throughout.

This looks fairly well and is great on the eyes
for reducing the large white pages that blinds you
in the night.

Change-Id: I01cf32f3decdb885307a76eb79d644667bbbf9a3