web: add theme selection and default to auto
This replaces the dark mode checkbox and instead makes
it a selection dropdown where you can select one of three
themes.

Auto will use your system preference to choose either the
light or dark theme, changes dynamically based on your
system preference.

Light is the current theme.

Dark is the theme added by this patch series.

The UX this changes is that if somebody has their system
preferences set to dark, for example in Mac OS X that is
in System Settings -> Appearance -> Dark the user will
get the Zuul web UI in dark by default and same for the
opposite.

Change-Id: Icfdfff3deb67bb706fa2d0d1068092e053f06e64