gerrit : filter unimplemented triggers
This was inspired after looking at the upgrade notes for Gerrit 3.7.
This version will output a new "project-head-updated" event in
response to set-head commands.  This new event is added to the mix
here.

At first I was wondering if Zuul might need to support triggering on
this event for some sort of workflow where maybe people want something
like post jobs to run when they switch a project HEAD.  Although the
event is new, the set-head idea isn't.  Thus I don't think it's
currently worth considerable effort to try and match this event when
we don't know if anyone does, or ever will, want to match it.

But, this did get me looking at all the other events as well.
Currently we only support matching on events listed in the schema in
driver/gerrit/gerrittrigger.py.  However, we let almost all events
through, even if they won't ever be matched.

The bubble through to the event filter matchers outputs multiple lines
of log info as it passes along; e.g. on the opendev.org instance,
using an old log file it's not massive as a % wise but it's still
3.2MiB of logs in the ~800MiB total log

 $ m=$(egrep '(reviewer-added|hashtags-changed|topic-changed|project-updated)' debug.log.1 | wc -c)
 $ t=$(cat debug.log.1 | wc -c)
 $ python3 -c "print($t/1024/1024)"
 821.254937171936
 $ python3 -c "print($m/1024/1024)"
 3.2903575897216797
 $ python3 -c "print($m/$t*100)"
 0.4006499615152777

This proposes that we add these unimplemented events to the ignore
list.

This does add a debug message, since it might be nice to know these
events are coming in but are being explicitly ignored.

Change-Id: I3e12b165a29a12a0a0ca6ea0940a7c888442c2ed