gerrit trigger : alphabetize event lists
This is a no-op refactor that alphabetizes the event lists.  Since we
keep the list in a few places (ignores, docs, and the author matcher)
it makes it easier to compare and make sure events line up if they're
in alphabetical order.

This also explicitly calls out in the trigger matcher that the docs
need updating.

Change-Id: I29a64f31de0235ab4be9aea7371088d4af65d41f