Fix Ansible python path detection
Ansible python path detection operates in two ways. First, if your
operating system is in INTERPRETER_PYTHON_DISTRO_MAP ansible looks up
the values there and is happy. If your distro is not in that mapping
(opensuse, oracle, arch etc) then Ansible uses
`command -v $pythonversion` to detect python for various versions within
your PATH.

This becomes a problem if you are running Zuul and Ansible out of a
virtualenv as that includes the virtualenv in PATH for local connection
execution of Ansible. This local connection of Ansible autodetects
python's path as being in that virtualenv. Then later [0] when Zuul runs
playbooks under bwrap the virtualenv is not mounted and python can not
be found leading to failure. Chances are low that this will ever be a
problem in production unless we have users running Zuul executors on
something other than Debuntu or Red Hat's family of distros and
derivatives. But I hit this running the test suite on my local OpenSUSE
Tumbleweed installations.

Ultimately, this feels like an Ansible bug applying inconsistent python
detection. However, it seems unlikely we'll remove the special casing
for popular distros, and we won't be able to list every distro in
Ansible's distro map. To fix this we refactor the execution context
wrapper libraries to track the list of bind mounted paths. Then we
filter any paths in PATH that do not get mounted. This ensures that the
null wrapper and bwrap wrapper run with a consistent PATH that both
sides are able to see. Ansible python path detection will no longer see
the virtualenv and can only find paths that are valid under bwrap.

To do this refactor I've also simplified the bwrap driver a bit moving
the handling of the bwrap command completely into the execution context
class and out of the wrapper class.

[0] The first playbook run is the setup playbook and this does not run
with the bwrap wrapper. The second playbook is the freeze playbook and
this does run under bwrap. The setup playbook finds the virtualenv
python and caches this for the freeze playbooks which fails.

Change-Id: I5c67af82a381e0840a297122daf62a867cd6776c