Only use "master" as default branch if not set
So far we've used "master" as default branch when we encounter a project
configuration that doesn't have a default branch set. The problem here
is, that later project stanzas that do define a default branch won't
have any effect.

To fix this, we will only fall back on using "master" in the project
metadata if none of the project definitions declares a default branch.

Change-Id: Iddd5ef906c2bdbcc472615ffe25033c4ca961df7