Increase Github event processor thread pool
So far we were using the default value for the thread pool executor
running the Github event pre-processing which is determined by
`min(32, (os.cpu_count() or 1) + 4)`.

However, on a busy system this value might be too small when a lot of
events arrive at the same time and some of the events take longer to
process. To avoid contention around the thread pool we'll change the
pool size to `min(32, (os.cpu_count() or 1) * 4)` (similar to the
default value before Python 3.8).

Change-Id: Ief349638fe41fabe7ca417f23bf439f860e524a7