gerrit: Handle case where commit has no parents
If a commit has no parents (e.g. the first commit in a repository), the
HTTP request to gerrit to the revision files endpoint must not specify
the parent=1 query parameter. Otherwise, gerrit will return an HTTP 400
error with the message 'invalid parent number: 1'.

Explicitly check for this by examining if any parents are present for
the current revision before adding the query parameter

Change-Id: I3f93899dc79ebc716223ecc6f5adecd0f1ce9b3e