Further fix getting topic changes by git needs
The test helper method that handles fake gerrit queries had a bug
which would cause the "topic:" queries to return all open changes.

When we correct that, we can see, by virtue of a newly raised
execption that there was some unexercised code in getChangesByTopic
which is now exercised.  This change also corrects the exception
that is raised when mutating a set while iterating over it.

Change-Id: I1874482b2c28fd1082fcd56036afb20333232409