Improve FrozenJob.isEqual
Some FrozenJob attributes did not compare cleanly because their
class implementations did not implement equality methods or the
in-memory attributes right after creation were not the same data
types as what was deserialized.  This could cause the job
deduplication to miss an opportunity to deduplicate a job.

This is corrected by implementing some equality and hash methods
along with some normalization of attributes passed to the
FrozenJob constructor.

Several deduplication tests are updated to force serialization
issues to appear by splitting the deduplication work across
multiple schedulers.

Change-Id: I7960aa2272ecfba90808c8b084614036ca0abc59