Fix JWT auth test and alembic migration
A new version of pyjwt was released which alters an internal call
it makes to urllib which we have mocked in the unit tests.  Our
mock must be updated to match.

The previous version would call urlopen with a string url argument,
newer versions call it with a Request object (urllib accepts both).

This change updates the mock to accept both.

There has also been a recent alembic release which alters the function
signature of alter_column so that most arguments are now required to
be keyword arguments.  This causes one of our migrations to fail
since it was not passing the new type in as a positional argument.

Notably, the type was not the next positional argument in the
signature ("nullable" is), so this explains why we had two
"change patchset to string" migrations: the first one did not actually
work.

This change alters the migration to do what it effectively did rather
that what it was intended to do.  The later migration continues to
correct the error.

Change-Id: I10cdffa43147f7f72e431e7c64e10e9416dfb295