Fix JWT auth test
A new version of pyjwt was released which alters an internal call
it makes to urllib which we have mocked in the unit tests.  Our
mock must be updated to match.

The previous version would call urlopen with a string url argument,
newer versions call it with a Request object (urllib accepts both).

This change updates the mock to accept both.

Change-Id: I10cdffa43147f7f72e431e7c64e10e9416dfb295