Set the start_time key for buildset events
The buildset event does not set the start_time attribute which makes the
index step fail. This commit gets the start_time attribute from the
current_buildset attribute and sets it in the event.

Finally, the commit also adds a check in the gen function to make sure the
event has a start_time key before accessing it. If it doesn't, it will set
the timestamp to the current utc time.

Fixes #2010763

Change-Id: I264b99e4403ca6cb6499307b84a023266063ea54
Signed-off-by: Flavio Percoco <flavio@pacerevenue.com>