Don't set buildset files for non-live items ahead
I82848367bd6f191ec5ae5822a1f438070cde14e1 introduced an optimization to
set the buildset files for non-live items if they don't already have
files set.

Later Id533772f35ebbc76910398e0e0fa50a3abfceb52 changed the approach to
also build a layout for non-live items with config updates.

Unfortunately the initial code path for setting the buildset files for
non-live items ahead was still in place.

In combination with circular config updates, this can cause problems
when merges are not completed in the same order as requested, which can
definitely happen in a production system. So far this issue went
unnoticed in the tests as the order there seems to be deterministic.

Change-Id: I38925e5fd0ed5ff45aab17d108740345716fd478